ERROR on line 535: setMediaSession (android.media.session.MediaSession.Token) in MediaStyle cannot be applied to (android.support.v4.media.session.MediaSessionCompat.Token)
Spent hours trying to figure this out. It seems to be caused by NotificationBuilder or some related class being deprecated. There does not appear to be an obvious workaround.

It pisses me off that when they break things they don't provide either a substitute that works in an identical manner or clear instructions on how to fix it.